requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1
Flask-WTF==1.1.1
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.4
Flask-Babel==3.0.0